* Fixed diagraphica warnings for GCC
git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@8554 21252fe5-60e8-0310-8ff0-c1ae1001594e

+++ tools/diagraphica/arcdiagram.h
@@ -159,7 +159,7 @@ protected:
     void handleHoverCluster(
         const size_t &i,
         const size_t &j );
-    void handleHoverBundle( const int &bndlIdx );
+    void handleHoverBundle( const size_t &bndlIdx );
     void handleHoverBarTree(
         const int &i,
         const int &j );
